Optimize DeclarativeEnvironmentRecord (#518)

+++ Jint/Runtime/TypeConverter.cs
@@ -266,7 +266,7 @@ namespace Jint.Runtime
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static string ToString(uint i)
         {
-            return i < intToString.Length
+            return i < (uint) intToString.Length
                 ? intToString[i]
                 : i.ToString();
         }
